Huge restructuring
Alright, pleae allow me to explain myself here.

It will be basically useless to look at the diffs on this commit because I restructured everything. The goal was to make this BSP a drop-in element. To use the BSP you need it to appear as a sub-directory of the "boards" folder in the SDK. There are two good ways to do that:
1) Clone, or download and extract, right into the boards folder of the SDK.
2) Clone into your favorite place to keep Git repos and then make a symbolic link into the boards folder of the SDK.

Within the repo all paths are relative so the actual name you use for this folder doen't matter but the location (in SDK/boards/) and the internal structure do matter.

Here's info on my changes:
- Removed 'fast_gpio' example for clarity in a tutorial. (May reconsider this)
- Renamed 'tensorflow_test' to 'edge_test' to reflect the updated board name
- Moved 'tools' folder underneath 'bsp' so that the SparkFun variant of the UART uploader script could be used without having to move it to another folder (the SDK/tools/ folder)
- Made a copy of 'am_defines.py' under 'tools/' because it was an easy short-term solution to make our UART uploader script work. This could be improved in the future. (Probably by passing in the path to the SDK/tools/ folder in our script and then importing their 'am_defines.py' file. I will also modify the name of our local 'am_defines.py' copy to 'sf_am_defines.py' as a hint
- Modified the makefile for the edge_test example as required to work with our script and all relative paths
- Created a template project that can be set up relatively easily from any system location