Add ADC DMA example
Initial commit of the ADC DMA example.

What you get with this example is the ADC sampling at 16 kHz on two channels (MIC0 and MIC1) using DMA while the uC is in deep sleep. If you have debugging turned on (default) then after each sample period the values are printed to UART in two columns (good for Arduino Serial Plotter) for visualization. B/c UART is so slow most of the time is spent in print statements, with short bursts of recording. Of course without printing recording is just limited by memory.

Learned a few things along the way:

* There is one ADC module, to get additonal channels you have to configure another 'slot'
* When configuring slots be sure to choose the correct slot number
* It appears that ctimer A3 has a special relationship with the ADC, being able to trigger it directly - may be that no other timers can do this
* Multi-channel samples are recorded all at once into the ADC's buffer. They need to be separated manually. The 'am_hal_adc_read_samples' function promises this capability in the doxygen markup but does not deliver. For now it is up to the user to do that (using a 'slot' member of each recording).